+++ lib/plausible_web/views/universal_analytics_view.ex
@@ -1,0 +1,4 @@
+defmodule PlausibleWeb.UniversalAnalyticsView do
+  use PlausibleWeb, :view
+  use Plausible
+end